urbvan/requirements/base.txt
Django==2.0.4
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.8.2
factory-boy==2.10.0
Faker==0.9.0
marshmallow==2.15.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
text-unidecode==1.2
